refactor: buttons to one button component
