+++ components/account/account.module.scss
@@ -62,7 +62,7 @@
   text-decoration: none;
   background-image: var(--gradient);
   padding: 0.5rem 2.5rem;
-  border-radius: 0.2em;
+  border-radius: var(--radius-small);
   font-weight: 500;
   cursor: pointer;
 }
