+++ components/common/input/input.module.scss
@@ -21,12 +21,9 @@
 
 .input {
   background: var(--bg-color);
-  width: 100%;
-  border-width: 2px;
-  border-style: solid;
-  border-radius: 0.25em;
-  border-color: rgb(0, 0, 0, 0.4);
-  box-shadow: var(--box-shadow);
+  border: 3px solid rgb(0, 0, 0, 0.4);
+  border-radius: 0.35em;
+  box-shadow: var(--box-shadow), var(--shadow-color) 0 0 0 2px;
   padding: 1.1em 0.35em 0.15em 0.35em;
   line-height: 1.5;
   outline: none;
@@ -68,9 +65,9 @@
   display: flex;
   gap: 0.5rem;
   left: 0.6em;
-  font-weight: 500;
   top: 0.8em;
   position: absolute;
+  font-weight: 500;
   transition: all var(--transition-fast);
   color: rgb(0, 0, 0, 0.6);
 
@@ -107,9 +104,16 @@
   font-family: inherit;
   border: none;
   background-color: transparent;
+  border: 4px solid transparent;
 
-  &:focus + .label {
-    @include label-active;
+  &:focus {
+    outline: none;
+    border-radius: var(--radius-small);
+    border-color: var(--primary-color);
+
+    & + .label {
+      @include label-active;
+    }
   }
 }
