+++ components/editAccount/buttons/buttons.module.scss
@@ -10,11 +10,11 @@
 
 .cancel,
 .submit {
-  padding: 0.75rem 1.5rem;
+  padding: 0.75rem 2rem;
   transition: opacity var(--transition-fast);
   box-shadow: var(--box-shadow);
   font-weight: 700;
-  border-radius: 0.25em;
+  border-radius: var(--radius-small);
   flex-grow: 0;
 
   &::first-letter {
