+++ styles/globals.scss
@@ -23,8 +23,11 @@
   --transition-fast: 150ms cubic-bezier(0.4, 0, 0.2, 1);
   --transition-slow: 300ms cubic-bezier(0.4, 0, 0.2, 1);
 
-  --box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px, rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
+  --box-shadow: rgba(50, 50, 93, 0.2) 0px 6px 12px -2px, rgba(0, 0, 0, 0.3) 0px 3px 5px -3px;
   --gray-border: 2px solid hsl(120, 0%, 70%);
+
+  --radius-small: 0.3em;
+  --radius-medium: 0.45em;
 }
 
 html {
